requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
phonenumbers==8.12.31
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-gettext==4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.6
